strcp_s instead of strcp non safe and val_blk editor correction

+++ src/DataContainers/lexerCsv.cpp
@@ -763,7 +763,7 @@ void LexerCsv::initialize()
 
 void LexerCsv::backward(QTextStream &in)
 {
-    int l = lexem.length() + 1;
+    qint64 l = lexem.length() + 1;
     in.seek(in.pos() - l);
 }
